// src/types/enums.ts
export enum Routes {
    BottomTabs = 'BottomTabs',
  
    // stacks
    HomeStack = 'Home Stack',
    NewsStack = 'News Stack',
    DealsStack = 'Deals Stack',
    PulseStack = 'Pulse Stack',
  
    // VIEWS
    //----Screens
    HomeScreen = 'Home Screen',
    NewsScreen = 'News Screen',
    DealsScreen = 'Deals Screen',
    PulseScreen = 'Pulse Screen',
  }
  
  export enum ICON_TYPE {
    SVG = 'SVG',
    IMAGE = 'IMAGE',
    OCTICONS = 'OCTICONS',
    IONICONS = 'IONICONS',
    ENTYPO = 'ENTYPO',
    FontAwesome = 'FontAwesome',
  }